Add single graph command
Also renamed add_graphs to add_graph_bulk. The difference between the two is
that the bulk add will fill in values for you based on the check search, while
the single one you fill them all in manually (or via a script) on the command
line. This makes it possible to do a single graph spanning multiple checks
using the single add graph command.